+++ app/javascript/components/Donation/Donation.jsx
@@ -3,14 +3,8 @@ import axios from "axios";
 import PaymentForm from "./PaymentForm";
 
 const Organizations = () => {
-  const INITIAL_STATE = {
-    firstName: "",
-    lastName: "",
-    donation: ""
-  };
   const [results, setResults] = useState(null);
   const [term, setTerm] = useState("");
-  const [formData, setFormData] = useState(INITIAL_STATE);
 
   useEffect(() => {
     fetchResults();
@@ -27,40 +21,24 @@ const Organizations = () => {
       });
   };
 
-  const handleInputChange = field => e => {
-    setFormData({ ...formData, [field]: e.target.value });
-    console.log(formData);
-  };
-
   const handleChange = event => {
     const fieldTerm = event.target.value.trim();
     setTerm(fieldTerm);
-    console.log(term);
   };
   return (
     <>
-      <PaymentForm />
+      <h1>Donate</h1>
       <div>
-        <label htmlFor="term">Search</label>
-        <br />
-        <input type="search" name="term" value={term} onChange={handleChange} />
-        <br />
-        <div>
-          <h3>Results</h3>
-          {results &&
-            results
-              .filter(result =>
-                result.name.toLowerCase().includes(term.toLowerCase())
-              )
-              .map((result, i) => {
-                return (
-                  <p key={i}>
-                    <h1>{result.name}</h1>
-                  </p>
-                );
-              })}
-        </div>
+        <p>
+          "No time to spare? You can still make a difference!
+          <br />
+          Donate any amount you are comfortable with to a partner organization
+          of your choice.
+          <br />
+          No gift is too small to make an impact on your community."
+        </p>
       </div>
+      <PaymentForm results={results} handleChange={handleChange} term={term} />
     </>
   );
 };
